third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEBlendElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEBlendElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS normalBlend.in1.baseVal is "SourceAlpha"
PASS multiplyBlend.in1.baseVal is "SourceAlpha"
PASS screenBlend.in1.baseVal is "SourceAlpha"
PASS darkenBlend.in1.baseVal is "SourceAlpha"
PASS lightenBlend.in1.baseVal is "SourceAlpha"
PASS normalBlend.in1.baseVal is "SourceGraphic"
PASS multiplyBlend.in1.baseVal is "SourceGraphic"
PASS screenBlend.in1.baseVal is "SourceGraphic"
PASS darkenBlend.in1.baseVal is "SourceGraphic"
PASS lightenBlend.in1.baseVal is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
